Revert "154 add chunks analysis part 2"
